--- alembic/versions/137bec66f203_add_issues_and_user.py
@@ -1,24 +1,0 @@
-"""add issues and user
-
-Revision ID: 137bec66f203
-Create Date: 2015-05-12 09:54:49.452506
-
-"""
-
-# revision identifiers, used by Alembic.
-revision = '137bec66f203'
-
-from alembic import op
-import sqlalchemy as sa
-
-
-def upgrade():
-    ### commands auto generated by Alembic - please adjust! ###
-    pass
-    ### end Alembic commands ###
-
-
-def downgrade():
-    ### commands auto generated by Alembic - please adjust! ###
-    pass
-    ### end Alembic commands ###
