+++ alembic/versions/2d876fb4ab3a_add_column_is_superuser_to_model.py
@@ -1,14 +1,12 @@
 """add column is_superuser to model
-
-Revision ID: "3ad1cf34f2f0"
-Revises: 3ad1cf34f2f0
+Revision ID: 2d876fb4ab3a
+Revises: 53c52e7884df
 Create Date: 2015-05-09 23:49:12.260119
-
 """
 
 # revision identifiers, used by Alembic.
-revision = '"3ad1cf34f2f0"'
-down_revision = '3ad1cf34f2f0'
+revision = '2d876fb4ab3a'
+down_revision = '53c52e7884df'
 
 from alembic import op
 import sqlalchemy as sa
