--- alembic/versions/528a609600b3_add_issues.py
@@ -1,26 +1,0 @@
-"""add issues
-
-Revision ID: 528a609600b3
-Revises: 4c28f392154d
-Create Date: 2015-05-12 09:46:36.031096
-
-"""
-
-# revision identifiers, used by Alembic.
-revision = '528a609600b3'
-down_revision = '137bec66f203'
-
-from alembic import op
-import sqlalchemy as sa
-
-
-def upgrade():
-    ### commands auto generated by Alembic - please adjust! ###
-    pass
-    ### end Alembic commands ###
-
-
-def downgrade():
-    ### commands auto generated by Alembic - please adjust! ###
-    ### end Alembic commands ###
-    pass
